Move config repos check connection to internal config repos api
