+++ api/api-config-repos-v4/src/test/groovy/com/thoughtworks/go/apiv4/configrepos/ConfigReposControllerV4Test.groovy
@@ -19,7 +19,6 @@ import com.thoughtworks.go.api.SecurityTestTrait
 import com.thoughtworks.go.api.spring.ApiAuthenticationHelper
 import com.thoughtworks.go.config.*
 import com.thoughtworks.go.config.materials.PasswordDeserializer
-import com.thoughtworks.go.config.materials.git.GitMaterial
 import com.thoughtworks.go.config.policy.Allow
 import com.thoughtworks.go.config.policy.Deny
 import com.thoughtworks.go.config.policy.Policy
@@ -29,20 +28,16 @@ import com.thoughtworks.go.config.remote.PartialConfig
 import com.thoughtworks.go.domain.PipelineGroups
 import com.thoughtworks.go.domain.materials.Material
 import com.thoughtworks.go.domain.materials.MaterialConfig
-import com.thoughtworks.go.domain.materials.ValidationBean
 import com.thoughtworks.go.server.domain.Username
 import com.thoughtworks.go.server.materials.MaterialUpdateService
 import com.thoughtworks.go.server.service.ConfigRepoService
 import com.thoughtworks.go.server.service.EntityHashingService
-import com.thoughtworks.go.server.service.GoConfigService
 import com.thoughtworks.go.server.service.MaterialConfigConverter
-import com.thoughtworks.go.server.service.SecretParamResolver
 import com.thoughtworks.go.server.service.result.HttpLocalizedOperationResult
 import com.thoughtworks.go.spark.ControllerTrait
 import com.thoughtworks.go.spark.NormalUserSecurity
 import com.thoughtworks.go.spark.Routes
 import com.thoughtworks.go.spark.SecurityServiceTrait
-import com.thoughtworks.go.util.SystemEnvironment
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
@@ -79,18 +74,6 @@ class ConfigReposControllerV4Test implements SecurityServiceTrait, ControllerTra
   @Mock
   private MaterialConfigConverter converter
 
-  @Mock
-  private PasswordDeserializer passwordDeserializer
-
-  @Mock
-  private MaterialConfigConverter materialConfigConverter
-
-  @Mock
-  private SystemEnvironment systemEnvironment
-
-  @Mock
-  private SecretParamResolver secretParamResolver
-
   @BeforeEach
   void setUp() {
     initMocks(this)
@@ -98,7 +81,7 @@ class ConfigReposControllerV4Test implements SecurityServiceTrait, ControllerTra
 
   @Override
   ConfigReposControllerV4 createControllerInstance() {
-    new ConfigReposControllerV4(new ApiAuthenticationHelper(securityService, goConfigService), goConfigService, passwordDeserializer, materialConfigConverter, systemEnvironment, secretParamResolver, service, entityHashingService, materialUpdateService, converter)
+    new ConfigReposControllerV4(new ApiAuthenticationHelper(securityService, goConfigService), service, entityHashingService, materialUpdateService, converter)
   }
 
   @Nested
@@ -222,18 +205,6 @@ class ConfigReposControllerV4Test implements SecurityServiceTrait, ControllerTra
       }
     }
 
-    @Nested
-    class CheckConnection implements SecurityTestTrait, NormalUserSecurity {
-      @Override
-      String getControllerMethodUnderTest() {
-          return "testConnection"
-      }
-
-      @Override
-      void makeHttpCall() {
-          postWithApiHeader(controller.controllerPath(ID_1, 'material_test'), [:])
-      }
-    }
   }
 
   @Nested
@@ -820,107 +791,6 @@ class ConfigReposControllerV4Test implements SecurityServiceTrait, ControllerTra
       assertThatResponse()
         .isForbidden()
         .hasJsonMessage("User '${currentUsername().getDisplayName()}' does not have permissions to administer 'test-id' config_repo(s).")
-    }
-  }
-
-  @Nested
-
-  class CheckConnection {
-    @BeforeEach
-    void setUp() {
-      loginAsAdmin()
-    }
-
-    @Test
-    void 'should render error if material type is invalid'() {
-      postWithApiHeader(controller.controllerPath(ID_1, 'material_test'), ["type": "some-random-type"])
-      assertThatResponse()
-              .isUnprocessableEntity()
-              .hasJsonMessage("Your request could not be processed. Invalid material type 'some-random-type'. It has to be one of [git, hg, svn, p4, tfs, dependency, package, plugin].")
-    }
-
-    @Test
-    void 'should render error if material type does not support check connection functionality'() {
-      postWithApiHeader(controller.controllerPath(ID_1, 'material_test'), ["type": "dependency"])
-      assertThatResponse()
-              .isUnprocessableEntity()
-              .hasJsonMessage("Your request could not be processed. The material of type 'dependency' does not support connection testing.")
-    }
-
-    @Test
-    void 'should render error if material is not valid'() {
-      postWithApiHeader(controller.controllerPath(ID_1, 'material_test'), [
-              "type"      : "git",
-              "attributes": [
-                      "url"               : "",
-                      "encrypted_password": "encrypted-password",
-                      "password"          : "password"
-              ]
-      ])
-
-      def expectedJSON = [
-              "message": "There was an error with the material configuration.\n- url: URL cannot be blank",
-              "data"   : [
-                      "errors"    : [
-                              "url": ["URL cannot be blank"]
-                      ],
-                      "type"      : "git",
-                      "attributes": [
-                              "url"             : "",
-                              "destination"     : null,
-                              "filter"          : null,
-                              "invert_filter"   : false,
-                              "name"            : null,
-                              "auto_update"     : true,
-                              "branch"          : "master",
-                              "submodule_folder": null,
-                              "shallow_clone"   : false
-                      ]
-              ]
-      ]
-
-      assertThatResponse()
-              .isUnprocessableEntity()
-              .hasJsonMessage("There was an error with the material configuration.\\n- url: URL cannot be blank")
-              .hasJsonBody(expectedJSON)
-    }
-
-    @Test
-    void 'should render error if cannot connect to material'() {
-      def material = mock(GitMaterial.class)
-      def validationBean = ValidationBean.notValid("some-error")
-      when(material.checkConnection(any())).thenReturn(validationBean)
-      when(materialConfigConverter.toMaterial(any(MaterialConfig.class))).thenReturn(material)
-
-      postWithApiHeader(controller.controllerPath(ID_1, 'material_test'), [
-              "type"      : "git",
-              "attributes": [
-                      "url": "some-url"
-              ]
-      ])
-
-      assertThatResponse()
-              .isUnprocessableEntity()
-              .hasJsonMessage("some-error")
-    }
-
-    @Test
-    void 'should render success response if can connect to material'() {
-      def material = mock(GitMaterial.class)
-      def validationBean = ValidationBean.valid()
-      when(material.checkConnection(any())).thenReturn(validationBean)
-      when(materialConfigConverter.toMaterial(any(MaterialConfig.class))).thenReturn(material)
-
-      postWithApiHeader(controller.controllerPath(ID_1, 'material_test'), [
-              "type"      : "git",
-              "attributes": [
-                      "url": "some-url"
-              ]
-      ])
-
-      assertThatResponse()
-              .isOk()
-              .hasJsonMessage("Connection OK.")
     }
   }
